Changed package name to com.appcoins.eskills2048.
