+++ 2048/2048/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.tpcstld.twozerogame">
+    package="com.appcoins.eskills2048">
 
     <uses-permission android:name="android.permission.INTERNET" />
