+++ 2048/2048/src/main/java/com/appcoins/eskills2048/model/RoomStatus.java
@@ -1,0 +1,5 @@
+package com.appcoins.eskills2048.model;
+
+public enum RoomStatus {
+  PENDING, COMPLETED
+}
