+++ 2048/2048/src/main/java/com/appcoins/eskills2048/usecase/SetScoreUseCase.java
@@ -1,0 +1,21 @@
+package com.appcoins.eskills2048.usecase;
+
+import com.appcoins.eskills2048.model.RoomResponse;
+import com.appcoins.eskills2048.model.UserStatus;
+import com.appcoins.eskills2048.repository.RoomRepository;
+import io.reactivex.Single;
+import io.reactivex.schedulers.Schedulers;
+
+public class SetScoreUseCase {
+
+  private final RoomRepository roomRepository;
+
+  public SetScoreUseCase(RoomRepository roomRepository) {
+    this.roomRepository = roomRepository;
+  }
+
+  public Single<RoomResponse> setScore(String session, long score) {
+    return roomRepository.patch(session, score, UserStatus.PLAYING)
+        .subscribeOn(Schedulers.io());
+  }
+}
