+++ 2048/2048/src/main/java/com/appcoins/eskills2048/vm/MainGameViewModel.java
@@ -1,0 +1,36 @@
+package com.appcoins.eskills2048.vm;
+
+import com.appcoins.eskills2048.model.RoomResponse;
+import com.appcoins.eskills2048.usecase.SetFinalScoreUseCase;
+import com.appcoins.eskills2048.usecase.SetScoreUseCase;
+import io.reactivex.Single;
+
+public class MainGameViewModel {
+
+  private final SetFinalScoreUseCase setFinalScoreUseCase;
+  private final MainGameViewModelData data;
+  private final SetScoreUseCase setScoreUseCase;
+
+  public MainGameViewModel(SetScoreUseCase setScoreUseCase, SetFinalScoreUseCase setFinalScoreUseCase,
+      MainGameViewModelData data) {
+    this.setScoreUseCase = setScoreUseCase;
+    this.setFinalScoreUseCase = setFinalScoreUseCase;
+    this.data = data;
+  }
+
+  public Single<RoomResponse> setFinalScore(long score) {
+    return setFinalScoreUseCase.setFinalScore(data.getSession(), score);
+  }
+
+  public Single<RoomResponse> setScore(long score) {
+    return setScoreUseCase.setScore(data.getSession(), score);
+  }
+
+  public String getWalletAddress() {
+    return data.getWalletAddress();
+  }
+
+  public String getSession() {
+    return data.getSession();
+  }
+}
