+++ 2048/2048/src/main/java/com/appcoins/eskills2048/vm/MainGameViewModelData.java
@@ -1,0 +1,17 @@
+package com.appcoins.eskills2048.vm;
+
+import lombok.Data;
+
+@Data
+public class MainGameViewModelData {
+
+  private final String userId;
+  private final String walletAddress;
+  private final String session;
+
+  public MainGameViewModelData(String userId, String walletAddress, String session) {
+    this.userId = userId;
+    this.walletAddress = walletAddress;
+    this.session = session;
+  }
+}
