--- 2048/2048/src/main/java/com/tpcstld/twozerogame/Tile.java
@@ -1,33 +1,0 @@
-package com.tpcstld.twozerogame;
-
-public class Tile extends Cell {
-    private final int value;
-    private Tile[] mergedFrom = null;
-
-    public Tile(int x, int y, int value) {
-        super(x, y);
-        this.value = value;
-    }
-
-    public Tile(Cell cell, int value) {
-        super(cell.getX(), cell.getY());
-        this.value = value;
-    }
-
-    public void updatePosition(Cell cell) {
-        this.setX(cell.getX());
-        this.setY(cell.getY());
-    }
-
-    public int getValue() {
-        return this.value;
-    }
-
-    public Tile[] getMergedFrom() {
-        return mergedFrom;
-    }
-
-    public void setMergedFrom(Tile[] tile) {
-        mergedFrom = tile;
-    }
-}
