--- 2048/2048/src/main/java/com/tpcstld/twozerogame/api/RoomApi.java
@@ -1,17 +1,0 @@
-package com.tpcstld.twozerogame.api;
-
-import com.tpcstld.twozerogame.model.PatchRoomRequest;
-import com.tpcstld.twozerogame.model.RoomResponse;
-import io.reactivex.Single;
-import retrofit2.http.Body;
-import retrofit2.http.GET;
-import retrofit2.http.Header;
-import retrofit2.http.PATCH;
-
-public interface RoomApi {
-
-  @PATCH("room/") Single<RoomResponse> patchRoom(@Header("authorization") String authorization,
-      @Body PatchRoomRequest patchRoomRequest);
-
-  @GET("room/") Single<RoomResponse> getRoom(@Header("authorization") String authorization);
-}
