--- 2048/2048/src/main/java/com/tpcstld/twozerogame/model/RoomStatus.java
@@ -1,5 +1,0 @@
-package com.tpcstld.twozerogame.model;
-
-public enum RoomStatus {
-  PENDING, COMPLETED
-}
