--- 2048/2048/src/main/java/com/tpcstld/twozerogame/model/UserStatus.java
@@ -1,5 +1,0 @@
-package com.tpcstld.twozerogame.model;
-
-public enum UserStatus {
-  PLAYING, COMPLETED
-}
